backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
anthropic>=0.50.0
pdfplumber==0.11.4
pandas==2.2.3
python-multipart==0.0.12
